Refactored sendImage
